Add PlayChordView
This is mostly a proof of concept, to be extended upon:
Extends the normal PlayView, but replaces top half of the grid with easy chord
buttons. These are derived from the currently selected scale and may not be
that useful for all scales.
Rows from top to bottom: 1-3-5-7, 1-3-5, 1st inversion, 2nd inversion.